Add lines for automatic goo updates pointing at goo.im/devs/beanstalk
I requested and was granted access for this goo account specifically
to use as a shared account for ALL of the builders/porters for
BeanStalk.  What you will need to do is get in touch with me
and i will provide the password and create a folder for your
device and let you know which one it is.  From there you just
have to upload your new builds every release and OTA should work fine

Signed-off-by: scott <scott@ubuntu.com>